# api-docs.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8090","description":"Generated server url"}],"paths":{"/profile":{"get":{"tags":["profile-controller"],"operationId":"listAllFormsOfMetadata_1","responses":{"200":{"description":"OK","content":{"application/hal+json":{"schema":{"$ref":"#/components/schemas/RepresentationModelObject"}}}}}}},"/profile/students":{"get":{"tags":["profile-controller"],"operationId":"descriptor_1_1_1","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}},"application/alps+json":{"schema":{"type":"string"}},"application/schema+json":{"schema":{"$ref":"#/components/schemas/JsonSchema"}}}}}}},"/students":{"get":{"tags":["JPA repository"],"description":"get-student","operationId":"getCollectionResource-student-get_1","parameters":[{"name":"page","in":"query","description":"Zero-based page index (0..N)","required":false,"schema":{"minimum":0,"type":"integer","default":0}},{"name":"size","in":"query","description":"The size of the page to be returned","required":false,"schema":{"minimum":1,"type":"integer","default":20}},{"name":"sort","in":"query","description":"Sorting criteria in the format: property,(asc|desc). Default sort order is ascending. Multiple sort criteria are supported.","required":false,"schema":{"type":"array","items":{"type":"string"}}}],"responses":{"200":{"description":"OK","content":{"application/hal+json":{"schema":{"$ref":"#/components/schemas/PagedModelEntityModelStudent"}},"application/x-spring-data-compact+json":{"schema":{"$ref":"#/components/schemas/PagedModelEntityModelStudent"}},"text/uri-list":{"schema":{"type":"string"}}}}}},"post":{"tags":["JPA repository"],"description":"create-student","operationId":"postCollectionResource-student-post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/StudentRequestBody"}}},"required":true},"responses":{"201":{"description":"Created","content":{"application/hal+json":{"schema":{"$ref":"#/components/schemas/EntityModelStudent"}}}}}}},"/students/search/findStudentByEmail":{"get":{"tags":["JPA repository"],"summary":"Query for finding students by his/her email address","operationId":"executeSearch-student-get","parameters":[{"name":"email","in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"application/hal+json":{"schema":{"$ref":"#/components/schemas/EntityModelStudent"}}}},"404":{"description":"Not Found"}}}},"/students/search/findStudentById":{"get":{"tags":["JPA repository"],"summary":"Query for finding student by his/her id","operationId":"executeSearch-student-get_1","parameters":[{"name":"studentId","in":"query","schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"application/hal+json":{"schema":{"$ref":"#/components/schemas/EntityModelStudent"}}}},"404":{"description":"Not Found"}}}},"/students/{id}":{"get":{"tags":["JPA repository"],"description":"get-student","operationId":"getItemResource-student-get","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"application/hal+json":{"schema":{"$ref":"#/components/schemas/EntityModelStudent"}}}},"404":{"description":"Not Found"}}},"put":{"tags":["JPA repository"],"description":"update-student","operationId":"putItemResource-student-put","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/StudentRequestBody"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/hal+json":{"schema":{"$ref":"#/components/schemas/EntityModelStudent"}}}},"201":{"description":"Created","content":{"application/hal+json":{"schema":{"$ref":"#/components/schemas/EntityModelStudent"}}}},"204":{"description":"No Content"}}},"delete":{"tags":["JPA repository"],"description":"delete-student","operationId":"deleteItemResource-student-delete","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"204":{"description":"No Content"},"404":{"description":"Not Found"}}},"patch":{"tags":["JPA repository"],"description":"patch-student","operationId":"patchItemResource-student-patch","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/StudentRequestBody"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/hal+json":{"schema":{"$ref":"#/components/schemas/EntityModelStudent"}}}},"204":{"description":"No Content"}}}},"/api/v1/student/{studentId}":{"put":{"tags":["Operations on students"],"summary":"Method for updating students","operationId":"updateStudent","parameters":[{"name":"studentId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"firstName","in":"query","required":false,"schema":{"type":"string"}},{"name":"lastName","in":"query","required":false,"schema":{"type":"string"}},{"name":"dateOfBirth","in":"query","required":false,"schema":{"type":"string"}},{"name":"email","in":"query","required":false,"schema":{"type":"string"}},{"name":"country","in":"query","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK"}}},"delete":{"tags":["Operations on students"],"summary":"Method for deleting student by his/her ID","operationId":"deleteStudent","parameters":[{"name":"studentId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/api/v1/student/addStudent":{"post":{"tags":["Operations on students"],"summary":"Method for adding students","operationId":"addStudent","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Student"}}},"required":true},"responses":{"200":{"description":"OK"}}}},"/api/v1/student":{"get":{"tags":["Operations on students"],"summary":"Method for getting all students in database","operationId":"getStudents","responses":{"200":{"description":"OK","content":{"application/hal+json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/StudentDTO"}}}}}}}},"/api/v1/student/{email}":{"get":{"tags":["Operations on students"],"summary":"Method for getting one student by his/her email address","operationId":"getStudentByEmail","parameters":[{"name":"email","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"application/hal+json":{"schema":{"$ref":"#/components/schemas/StudentDTO"}}}}}}},"/api/v1/student/regionsByCountry/{studentId}":{"get":{"tags":["Operations on students"],"summary":"Method for getting students' regions","operationId":"getRegionsByStudentId","parameters":[{"name":"studentId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"application/hal+json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/RegionAndSubregionDTO"}}}}}}}},"/api/v1/student/id/{id}":{"get":{"tags":["Operations on students"],"summary":"Method for getting student's country by his/her id","operationId":"getCountryById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"application/hal+json":{"schema":{"type":"string"}}}}}}}},"components":{"schemas":{"AbstractJsonSchemaPropertyObject":{"type":"object","properties":{"title":{"type":"string"},"readOnly":{"type":"boolean"}}},"Item":{"type":"object","properties":{"type":{"type":"string"},"properties":{"type":"object","additionalProperties":{"$ref":"#/components/schemas/AbstractJsonSchemaPropertyObject"}},"requiredProperties":{"type":"array","items":{"type":"string"}}}},"JsonSchema":{"type":"object","properties":{"title":{"type":"string"},"description":{"type":"string"},"properties":{"type":"object","additionalProperties":{"$ref":"#/components/schemas/AbstractJsonSchemaPropertyObject"}},"requiredProperties":{"type":"array","items":{"type":"string"}},"definitions":{"type":"object","additionalProperties":{"$ref":"#/components/schemas/Item"}},"type":{"type":"string"},"$schema":{"type":"string"}}},"Links":{"type":"object","additionalProperties":{"$ref":"#/components/schemas/Link"}},"RepresentationModelObject":{"type":"object","properties":{"_links":{"$ref":"#/components/schemas/Links"}}},"Student":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"firstName":{"type":"string"},"lastName":{"type":"string"},"email":{"type":"string"},"dateOfBirth":{"type":"string","format":"date"},"country":{"type":"string"},"age":{"type":"integer","format":"int32"}}},"EntityModelStudent":{"type":"object","properties":{"firstName":{"type":"string"},"lastName":{"type":"string"},"email":{"type":"string"},"dateOfBirth":{"type":"string","format":"date"},"country":{"type":"string"},"age":{"type":"integer","format":"int32"},"_links":{"$ref":"#/components/schemas/Links"}}},"PageMetadata":{"type":"object","properties":{"size":{"type":"integer","format":"int64"},"totalElements":{"type":"integer","format":"int64"},"totalPages":{"type":"integer","format":"int64"},"number":{"type":"integer","format":"int64"}}},"PagedModelEntityModelStudent":{"type":"object","properties":{"_embedded":{"type":"object","properties":{"students":{"type":"array","items":{"$ref":"#/components/schemas/EntityModelStudent"}}}},"_links":{"$ref":"#/components/schemas/Links"},"page":{"$ref":"#/components/schemas/PageMetadata"}}},"StudentRequestBody":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"firstName":{"type":"string"},"lastName":{"type":"string"},"email":{"type":"string"},"dateOfBirth":{"type":"string","format":"date"},"country":{"type":"string"},"age":{"type":"integer","format":"int32"}}},"StudentDTO":{"type":"object","properties":{"firstName":{"type":"string"},"lastName":{"type":"string"},"email":{"type":"string"},"country":{"type":"string"},"age":{"type":"integer","format":"int32"}}},"RegionAndSubregionDTO":{"type":"object","properties":{"region":{"type":"string"},"subregion":{"type":"string"}}},"Link":{"type":"object","properties":{"href":{"type":"string"},"hreflang":{"type":"string"},"title":{"type":"string"},"type":{"type":"string"},"deprecation":{"type":"string"},"profile":{"type":"string"},"name":{"type":"string"},"templated":{"type":"boolean"}}}}}}